requirements.txt
fastapi 
uvicorn 
python-jose[cryptography] 
requests
cryptography
python-jose
pyjwt
cryptography
jwcrypto
sqlmodel
duckdb-engine
duckdb
pytest
httpx